quality

+++ backend/src/constants.rs
@@ -15,8 +15,8 @@ pub const SKILL_COST_INCREASE_FACTOR: f64 = 0.31;
 pub const XP_INCREASE_FACTOR: f64 = 0.4;
 pub const ARMOR_FACTOR: f64 = 100.0;
 
-pub const MAX_ITEM_QUALITY_PER_LEVEL: f64 = 0.5;
-pub const MAX_ITEM_QUALITY: f64 = 25.0;
+pub const MAX_ITEM_QUALITY_PER_LEVEL: f32 = 0.5;
+pub const MAX_ITEM_QUALITY: f32 = 25.0;
 
 pub const MAX_MARKET_PUBLIC_LISTINGS: i64 = 100;
 pub const MAX_MARKET_PRIVATE_LISTINGS: i64 = 10;
